site: quicker redirects to appease the SEO deities
