+++ sites/svelte.dev/src/routes/docs.html/+server.js
@@ -1,0 +1,2 @@
+export const prerender = true;
+export { GET } from '../docs/+server.js';
