+++ sites/svelte.dev/src/routes/docs/+server.js
@@ -1,0 +1,12 @@
+import js from './redirect.js?raw';
+
+// prerenderer will choke otherwise
+export const prerender = false;
+
+export function GET() {
+	return new Response(`<html><head><script>${js}</script></head></html>`, {
+		headers: {
+			'content-type': 'text/html'
+		}
+	});
+}
